config: keep the special-file test off Windows at build time
syscall.Mkfifo does not exist there, so a runtime GOOS skip still fails
to compile. Moved behind a unix build tag. Checked with GOOS=windows,
darwin and linux vet.
